+++ NEWS.md
@@ -6,6 +6,7 @@ Updates
 * Simplifies end-user functions a bit (i.e. removes `include_groups` option)
 * Optimizes the code a bit, specially at the joins with tables in the package
 * Fixes codes duplication when both product and group/community match for a search
+* Includes both official and shortened section names
 
 # Version 0.3.1